refactor(vcs): move Git source fetchers into adapter
