+++ src/core/vcs/git.ts
@@ -6,12 +6,25 @@ import {
   buildGitShowArgs,
   buildGitStashShowArgs,
   listGitUntrackedFiles,
+  resolveGitCommitRef,
+  resolveGitDiffEndpoints,
   resolveGitRepoRoot,
   runGitText,
+  type GitDiffEndpoint,
+  type GitDiffEndpoints,
 } from "../git";
-import { createSkippedLargeMetadata } from "../diffFile";
+import {
+  createSkippedLargeMetadata,
+  type BuildDiffFileOptions,
+  type DiffFileSourceContext,
+} from "../diffFile";
+import {
+  createFileSourceFetcher,
+  type FileSourceFetcherOptions,
+  type FileSourceSpec,
+} from "../fileSource";
 import type { DiffFile } from "../types";
-import type { VcsAdapter } from "./types";
+import type { VcsAdapter, VcsReviewOperation } from "./types";
 
 const LARGE_DIFF_FILE_MAX_BYTES = 1_000_000;
 const LARGE_DIFF_FILE_MAX_LINES = 20_000;
@@ -93,6 +106,111 @@ function buildSkippedLargeTrackedDiffFile(
   };
 }
 
+interface ResolvedFileSourceSpecs {
+  old: FileSourceSpec;
+  new: FileSourceSpec;
+}
+
+/** Build a binary-aware source-fetcher factory from per-file source specs. */
+function createSourceFetcherBuilder(
+  resolveSpecs: (file: DiffFileSourceContext) => ResolvedFileSourceSpecs | undefined,
+  options: FileSourceFetcherOptions = {},
+): NonNullable<BuildDiffFileOptions["sourceFetcherBuilder"]> {
+  return (file) => {
+    if (file.isBinary) {
+      return undefined;
+    }
+
+    const specs = resolveSpecs(file);
+    return specs ? createFileSourceFetcher(specs, options) : undefined;
+  };
+}
+
+/** Convert one Git diff endpoint into the corresponding source lookup. */
+function gitEndpointSourceSpec(
+  endpoint: GitDiffEndpoint,
+  repoRoot: string,
+  filePath: string,
+): FileSourceSpec {
+  switch (endpoint.kind) {
+    case "none":
+      return { kind: "none" };
+    case "git-ref":
+      return { kind: "git-blob", repoRoot, ref: endpoint.ref, path: filePath };
+    case "index":
+      return { kind: "git-index", repoRoot, path: filePath };
+    case "worktree":
+      return { kind: "fs", absolutePath: join(repoRoot, filePath) };
+  }
+}
+
+/** Build source fetchers from exact Git old/new endpoints. */
+function buildGitEndpointSourceFetcherBuilder(
+  repoRoot: string,
+  endpoints: GitDiffEndpoints,
+  options: FileSourceFetcherOptions = {},
+): NonNullable<BuildDiffFileOptions["sourceFetcherBuilder"]> {
+  return createSourceFetcherBuilder(({ path, previousPath, type }) => {
+    const oldPath = previousPath ?? path;
+
+    return {
+      old:
+        type === "new" ? { kind: "none" } : gitEndpointSourceSpec(endpoints.old, repoRoot, oldPath),
+      new:
+        type === "deleted"
+          ? { kind: "none" }
+          : gitEndpointSourceSpec(endpoints.new, repoRoot, path),
+    };
+  }, options);
+}
+
+function buildRefRangeSourceFetcherBuilder(
+  repoRoot: string,
+  oldRef: string,
+  newRef: string,
+  options: FileSourceFetcherOptions = {},
+): NonNullable<BuildDiffFileOptions["sourceFetcherBuilder"]> {
+  return buildGitEndpointSourceFetcherBuilder(
+    repoRoot,
+    {
+      old: { kind: "git-ref", ref: oldRef },
+      new: { kind: "git-ref", ref: newRef },
+    },
+    options,
+  );
+}
+
+/** Build source fetchers for Git review operations when both source sides are exact. */
+function buildGitReviewSourceFetcherBuilder(
+  operation: VcsReviewOperation,
+  repoRoot: string,
+  cwd: string,
+  gitExecutable = "git",
+): NonNullable<BuildDiffFileOptions["sourceFetcherBuilder"]> | undefined {
+  switch (operation.kind) {
+    case "working-tree-diff": {
+      const endpoints = resolveGitDiffEndpoints(operation.input, { cwd, repoRoot, gitExecutable });
+      return endpoints
+        ? buildGitEndpointSourceFetcherBuilder(repoRoot, endpoints, { gitExecutable })
+        : undefined;
+    }
+    case "revision-show": {
+      const newRef = resolveGitCommitRef(operation.input, operation.input.ref ?? "HEAD", {
+        cwd: repoRoot,
+        gitExecutable,
+      });
+      return buildRefRangeSourceFetcherBuilder(repoRoot, `${newRef}^`, newRef, { gitExecutable });
+    }
+    case "stash-show": {
+      const newRef = resolveGitCommitRef(operation.input, operation.input.ref ?? "stash@{0}", {
+        cwd: repoRoot,
+        gitExecutable,
+      });
+      return buildRefRangeSourceFetcherBuilder(repoRoot, `${newRef}^`, newRef, { gitExecutable });
+    }
+  }
+}
+
 /** VCS adapter translating neutral review operations to Git commands. */
 export const gitAdapter: VcsAdapter = {
   id: "git",
@@ -100,6 +218,7 @@ export const gitAdapter: VcsAdapter = {
   capabilities: {
     reviewOperations: new Set(["working-tree-diff", "revision-show", "stash-show"]),
     stagedDiff: true,
+    sourceFetching: true,
     watchSignatures: true,
   },
 
@@ -132,6 +251,12 @@ export const gitAdapter: VcsAdapter = {
             cwd,
             gitExecutable,
           }),
+          sourceFetcherBuilder: buildGitReviewSourceFetcherBuilder(
+            operation,
+            repoRoot,
+            cwd,
+            gitExecutable,
+          ),
           untrackedFiles: listGitUntrackedFiles(input, { cwd, repoRoot, gitExecutable }),
           extraFiles: largeTrackedFiles.map((file, index) =>
             buildSkippedLargeTrackedDiffFile(file, index, repoRoot),
@@ -147,6 +272,12 @@ export const gitAdapter: VcsAdapter = {
           sourceLabel: repoRoot,
           title: input.ref ? `${repoName} show ${input.ref}` : `${repoName} show HEAD`,
           patchText: runGitText({ input, args: buildGitShowArgs(input), cwd, gitExecutable }),
+          sourceFetcherBuilder: buildGitReviewSourceFetcherBuilder(
+            operation,
+            repoRoot,
+            cwd,
+            gitExecutable,
+          ),
         };
       }
       case "stash-show": {
@@ -158,6 +289,12 @@ export const gitAdapter: VcsAdapter = {
           sourceLabel: repoRoot,
           title: input.ref ? `${repoName} stash ${input.ref}` : `${repoName} stash`,
           patchText: runGitText({ input, args: buildGitStashShowArgs(input), cwd, gitExecutable }),
+          sourceFetcherBuilder: buildGitReviewSourceFetcherBuilder(
+            operation,
+            repoRoot,
+            cwd,
+            gitExecutable,
+          ),
         };
       }
     }
